requirements.txt
-e .
aws-cdk.aws-kendra==1.127.0
aws-cdk.aws-iam==1.127.0